move files to make one library only
